% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_params.R
\name{theta2theta}
\alias{theta2theta}
\alias{set_thetas_equal}
\title{Set Process Probabilities Equal}
\usage{
theta2theta(model, names)

set_thetas_equal(model, names)
}
\arguments{
\item{model}{A list of the class \code{rtmpt_model}.}

\item{names}{Character vector giving the names of the processes for which the process probabilities should be equal. If 
\code{length(names) = 1} then the corresponding process probability will be estimates (i.e., it will be set to NA)}
}
\value{
A list of the class \code{rtmpt_model}.
}
\description{
Setting multiple process probabilities (thetas) equal. One of the process probabilities will be estimated and
  the other named process(es) will be set to equal the former. The equality can be removed by only using one name of a process.
}
\note{
If you use \code{theta2theta()} and \code{tau2tau()} with the same process names you might just change the EQN or MDL file accordingly
  by using the same process name for all processes which should have equal process times and probabilities.
}
\examples{
####################################################################################
# Detect-Guess variant of the Two-High Threshold model.
# The encoding and motor execution times are assumed to be equal for each category.
# The process probabilities for both detection processes ("do" and "dn") will be
# set equal.
####################################################################################

mdl_2HTM <- "
# targets
do+(1-do)*g
(1-do)*(1-g)

# lures
(1-dn)*g
dn+(1-dn)*(1-g)

# do: detect old; dn: detect new; g: guess
"

model <- to_rtmpt_model(mdl_file = mdl_2HTM)

## make do = dn
model <- theta2theta(model = model, names = c("do", "dn"))
model

####################################################################################
# Detect-Guess variant of the Two-High Threshold model.
# The encoding and motor execution times are assumed to be equal for each category.
# The process probabilities for both detection processes ("do" and "dn") will be
# set equal.
####################################################################################

mdl_2HTM <- "
# targets
do+(1-do)*g
(1-do)*(1-g)

# lures
(1-dn)*g
dn+(1-dn)*(1-g)

# do: detect old; dn: detect new; g: guess
"

model <- to_rtmpt_model(mdl_file = mdl_2HTM)

## make do = dn
model <- set_thetas_equal(model = model, names = c("do", "dn"))
model
}
\seealso{
\code{\link{set_resps}}, \code{\link{theta2const}}, \code{\link{tau2zero}} and \code{\link{tau2tau}}
}
\author{
Raphael Hartmann
}
