% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rtmpt.R
\name{update_drtmpt}
\alias{update_drtmpt}
\title{Update Diffusion RT-MPT Models}
\usage{
update_drtmpt(
  fit,
  data,
  n.iter = 1000,
  n.thin = 1,
  Irep = 1000,
  flags = NULL,
  control = NULL
)
}
\arguments{
\item{fit}{A list of the class \code{drtmpt_model}.}

\item{data}{Optimally, a list of class \code{drtmpt_data}. Also possible is a \code{data.frame} or a
path to the text file. Both, \code{data.frame} and the text file must contain the column names "subj",
"group", "tree", "cat", and "rt" preferably but not necessarily in this order. The values of the latter must
be in milliseconds. It is always advised to use \code{\link{to_drtmpt_data}} first, which gives back a
\code{drtmpt_data} list with information about the changes in the data, that were needed.}

\item{n.iter}{Number of samples per chain. Default is 1000.}

\item{n.thin}{Thinning factor. Default is 1.}

\item{Irep}{Every \code{Irep} samples an interim state with the current maximal potential scale reduction
factor is shown. Default is 1000. The following statements must hold true for \code{Irep}:
\itemize{
  \item \code{Irep} is a multiple of \code{n.thin} and
  \item \code{n.iter} is a multiple of \code{Irep / n.thin}.
}}

\item{flags}{Either NULL or a list of
\itemize{
  \item \code{old_label} If set to \code{TRUE} the old labels of "subj" and "group" of the data will be used in the elements of the output list.
    Default is \code{FALSE}.
  \item \code{indices} Model selection indices. If set to \code{TRUE} the log-likelihood for each iteration and trial will be
    stored temporarily and with that the DIC, as well as the WAIC and LOOIC will be calculated via the \code{loo} package. If you want to have the
    log-likelihood matrix stored in the output of this function, you can set \code{loglik} to \code{TRUE}. Default for
    \code{indices} is \code{FALSE}.
  \item \code{loglik} If set to \code{TRUE} and \code{indices = TRUE} the log-likelihood matrix for each iteration and trial will
    be saved in the output as a matrix. Default is \code{FALSE}.
}}

\item{control}{Either NULL or a list of
\itemize{
  \item \code{maxthreads} for the ML estimation of the initial values and the calculation of the DIC values one can use more than
    \code{n.chains} threads for parallelization. Default is 4 like \code{n.chians}. \code{maxthreads} must be larger or equal to
    \code{n.chains}.
}}
}
\value{
A list of the class \code{drtmpt_fit} containing
  \itemize{
    \item \code{samples}: the posterior samples as an \code{mcmc.list} object,
    \item \code{diags}: some diagnostics like deviance information criterion, posterior predictive checks for the frequencies and latencies,
                        potential scale reduction factors, and also the 99\% and 95\% HDIs and medians for the group-level parameters,
    \item \code{specs}: some model specifications like the model, arguments of the model call, and information about the data transformation,
    \item \code{indices} (optional): if enabled, WAIC and LOO,
    \item \code{LogLik} (optional): if enabled, the log-likelihood matrix used for WAIC and LOO.
    \item \code{summary} includes posterior mean and median of the main parameters.
  }
}
\description{
Given a diffusion RT-MPT fit object and the data, this function updates the fit. All information is taken from the fit object, except
  the number of iterations and thinning as well as some flags and control options.
}
\examples{
####################################################################################
# Detect-Guess variant of the Two-High Threshold model.
# The encoding and motor execution times are assumed to be equal for each response.
####################################################################################

mdl_2HTM <- "
# targets
do+(1-do)*g
(1-do)*(1-g)

# lures
(1-dn)*g
dn+(1-dn)*(1-g)

# do: detect old; dn: detect new; g: guess
"

model <- to_drtmpt_model(mdl_file = mdl_2HTM)

data_file <- system.file("extdata/data.txt", package="rtmpt")
data <- read.table(file = data_file, header = TRUE)
data_list <- to_drtmpt_data(raw_data = data, model = model)
\donttest{
# This might take some time
drtmpt_out <- fit_drtmpt(model = model, data = data_list, Rhat_max = 1.1)

drtmpt_updt <- update_drtmpt(fit = drtmpt_out, data = data_list, 
                             n.iter = 2000, n.thin = 1, Irep = 2000)
}
}
\references{
Klauer, K. C. (2010). Hierarchical multinomial processing tree models: A latent-trait approach. \emph{Psychometrika, 75(1)}, 70-98.

Spiegelhalter, D. J., Best, N. G., Carlin, B. P., & Van Der Linde, A. (2002). Bayesian measures of model complexity and fit.
  \emph{Journal of the royal statistical society: Series b (statistical methodology), 64(4)}, 583-639.

Vehtari, A., Gelman, A., & Gabry, J. (2017). Practical Bayesian model evaluation using leave-one-out cross-validation and WAIC.
  \emph{Statistics and Computing, 27(5)}, 1413-1432.

Watanabe, S. (2010). Asymptotic equivalence of Bayes cross validation and widely applicable information criterion in singular learning theory.
  \emph{Journal of Machine Learning Research, 11(Dec)}, 3571-3594.
}
\author{
Raphael Hartmann
}
